Renderer an Camera Rework Pt. 1
Started reworking the rendering logic.
All renderable components will now be passed to the project's renderer which will handle what objects are to be renderered to a camera view.
Also cameras will now only render/draw the elements that are actually in view and save the current view in its own canvas.

Still needs polishing and finishing touches.